Make DatumGetPointer() and PointerGetDatum() macros.
These functions as inline functions were producing warnings, based on discussion
https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/6fgKvN9QpV4/zjysjqIZAgAJ
converting them to macro as upstream. Adding explicit type casting wherever
needed now that DatumGetPointer() returns (char *) instead of (void *).

+++ src/backend/cdb/cdbhash.c
@@ -311,7 +311,7 @@ hashDatum(Datum datum, Oid type, datumHashFunction hashFn, void *clientData)
              * If we did a pg_detoast_datum, we need to remember to pfree, 
              * or we will leak memory.  Because of the 1-byte varlena header stuff.
              */
-            if (num != DatumGetPointer(datum)) 
+            if (num != (Numeric) DatumGetPointer(datum)) 
                 tofree = num;
 
 			break;
